backend/requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
billiard==4.2.4
ccxt==4.5.29
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coincurve==21.0.0
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.127.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
kombu==5.6.1
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
